pyproject.toml
[project]
name = "nac-fields"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "open3d>=0.19.0",
    "pymcubes>=0.1.6",
    "scikit-image>=0.25.2",
    "torch>=2.8.0",
    "trimesh>=4.7.4",
    "numpy==2",
    "numba>=0.61.2",
]
